+++ src/constants/settings.ts
@@ -71,6 +71,7 @@ export const settings = {
   trayIcon: "trayIcon",
   trayIconPath: "trayIconPath",
   updateFrequency: "updateFrequency",
+  windowTransparency: "windowTransparency",
   windowBounds: {
     root: "windowBounds",
     width: "windowBounds.width",